+++ packages/expo-router/src/fork/native-stack/NativeStackView.tsx
@@ -21,13 +21,13 @@ function NativeStackViewInner(props: ComponentProps<typeof RNNativeStackView>) {
       return props.state;
     }
 
-    const activeRoutes = props.state.routes;
-    const preloadedRoutes = props.state.preloadedRoutes;
+    const activeRoutes = props.state.routes.slice(0, props.state.index + 1);
+    const preloadedRoutes = props.state.routes.slice(props.state.index + 1);
 
     return {
       ...props.state,
-      routes: activeRoutes.concat(rootModals.routes),
-      preloadedRoutes,
+      index: props.state.index + rootModals.routes.length,
+      routes: activeRoutes.concat(rootModals.routes, preloadedRoutes),
     };
   }, [props.state, rootModals.routes]);