+++ packages/expo-router/src/global-state/router.ts
@@ -10,7 +10,6 @@ import {
   emitDomSetParams,
 } from '../domComponents/emitDomEvent';
 import { resolveHref } from '../link/href';
-import type { NavigationState } from '../react-navigation/native';
 import type { Href, RoutePath, RouteInputParams } from '../types';
 import { getHistoryLength } from '../utils/stack';
 import { shouldLinkExternally } from '../utils/url';
@@ -93,7 +92,7 @@ export function canDismiss(): boolean {
 
   // Keep traversing down the state tree until we find a stack navigator that we can pop
   while (state) {
-    if (state.type === 'stack' && getHistoryLength(state as NavigationState) > 1) {
+    if (state.type === 'stack' && getHistoryLength(state) > 1) {
       return true;
     }
     if (state.index === undefined) return false;