+++ packages/expo-router/src/modal/web/ModalStack.tsx
@@ -37,7 +37,7 @@ function ModalStackNavigator({
   children,
   screenOptions,
 }: ModalStackNavigatorProps) {
-  const { state, navigation, descriptors, NavigationContent, describe } = useNavigationBuilder<
+  const { state, navigation, descriptors, NavigationContent } = useNavigationBuilder<
     StackNavigationState<ParamListBase>,
     StackRouterOptions,
     StackActionHelpers<ParamListBase>,
@@ -73,21 +73,16 @@ function ModalStackNavigator({
 
   return (
     <NavigationContent>
-      <ModalStackView
-        state={state}
-        navigation={navigation}
-        descriptors={descriptors}
-        describe={describe}
-      />
+      <ModalStackView state={state} navigation={navigation} descriptors={descriptors} />
     </NavigationContent>
   );
 }
 
-const ModalStackView = ({ state, navigation, descriptors, describe }: ModalStackViewProps) => {
+const ModalStackView = ({ state, navigation, descriptors }: ModalStackViewProps) => {
   const isWeb = process.env.EXPO_OS === 'web';
   const { colors } = useTheme();
   const { preventedRoutes } = usePreventRemoveContext();
-  const activeRoutes = state.routes;
+  const activeRoutes = state.routes.slice(0, state.index + 1);
 
   const { routes: filteredRoutes, index: nonModalIndex } = convertStackStateToNonModalState(
     state,
@@ -115,13 +110,14 @@ const ModalStackView = ({ state, navigation, descriptors, describe }: ModalStack
 
   return (
     <div style={{ flex: 1, display: 'flex' }}>
-      <NativeStackView
-        state={newStackState}
-        descriptors={descriptors}
-        describe={describe}
-        emit={navigation.emit}
-        pop={pop}
-      />
+      {newStackState.routes.length > 0 && (
+        <NativeStackView
+          state={newStackState}
+          descriptors={descriptors}
+          emit={navigation.emit}
+          pop={pop}
+        />
+      )}
       {isWeb &&
         overlayRoutes.map((route) => {
           const isTransparentModal = isTransparentModalPresentation(