+++ packages/expo-router/src/modal/web/utils.ts
@@ -76,10 +76,11 @@ export function convertStackStateToNonModalState(
   descriptors: Record<string, { options: ExtendedStackNavigationOptions }>,
   isWeb: boolean
 ) {
-  const activeRoutes = state.routes;
+  const activeRoutes = state.routes.slice(0, state.index + 1);
+  const preloadedRoutes = state.routes.slice(state.index + 1);
 
   if (!isWeb) {
-    return { routes: activeRoutes, index: state.index };
+    return { routes: state.routes, index: state.index };
   }
 
   // Remove every modal-type route from the stack on web.
@@ -91,10 +92,11 @@ export function convertStackStateToNonModalState(
   // if that route was filtered out – at the last remaining route.
   let index = routes.findIndex((r) => r.key === activeRoutes[state.index]?.key);
   if (index < 0) {
-    index = routes.length > 0 ? routes.length - 1 : 0;
+    index = routes.length - 1;
   }
 
-  return { routes, index };
+  // Keep preloaded routes after the recalculated active index. If there are no active routes, index will be -1.
+  return { routes: routes.concat(preloadedRoutes), index };
 }
 
 /**
@@ -110,7 +112,7 @@ export function findLastNonModalIndex(
   state: StackNavigationState<ParamListBase>,
   descriptors: Record<string, { options: ExtendedStackNavigationOptions }>
 ) {
-  const activeRoutes = state.routes;
+  const activeRoutes = state.routes.slice(0, state.index + 1);
   // Iterate backwards through the stack to find the last non-modal route.
   for (let i = activeRoutes.length - 1; i >= 0; i--) {
     if (!isModalPresentation(descriptors[activeRoutes[i]!.key]!.options)) {