+++ packages/expo-router/src/react-navigation/core/useNavigationBuilder.tsx
@@ -896,7 +896,7 @@ export function useNavigationBuilder<
     getStateListeners: keyedListeners.getState,
   });
 
-  const { describe, descriptors } = useDescriptors<State, ActionHelpers, ScreenOptions, EventMap>({
+  const descriptors = useDescriptors<State, ActionHelpers, ScreenOptions, EventMap>({
     routes: state.routes,
     screens,
     navigation,
@@ -946,7 +946,6 @@ export function useNavigationBuilder<
   return {
     state,
     navigation,
-    describe,
     descriptors,
     NavigationContent,
   };