+++ packages/expo-router/src/react-navigation/native-stack/navigators/createNativeStackNavigator.tsx
@@ -38,7 +38,7 @@ function NativeStackNavigator({
   UNSTABLE_router,
   ...rest
 }: NativeStackNavigatorProps) {
-  const { state, describe, descriptors, navigation, NavigationContent } = useNavigationBuilder<
+  const { state, descriptors, navigation, NavigationContent } = useNavigationBuilder<
     StackNavigationState<ParamListBase>,
     StackRouterOptions,
     StackActionHelpers<ParamListBase>,
@@ -92,7 +92,6 @@ function NativeStackNavigator({
         {...rest}
         state={state}
         descriptors={descriptors}
-        describe={describe}
         emit={navigation.emit}
         pop={pop}
       />