+++ packages/expo-router/src/utils/stack.ts
@@ -1,12 +1,18 @@
-import type {
-  NavigationState,
-  ParamListBase,
-  StackNavigationState,
-} from '../react-navigation/native';
+import type { ReactNavigationState } from '../global-state/types';
+import type { NavigationState } from '../react-navigation/native';
 
-export function getHistoryLength(state: NavigationState): number {
+export function getHistoryLength(state: ReactNavigationState): number {
   if (state.history) {
     return state.history.length;
+  }
+
+  if (state.type === 'stack') {
+    if (state.index === undefined) {
+      return 1;
+    }
+
+    // All routes after `state.index` are preloaded.
+    return state.index + 1;
   }
 
   return state.routes.length;
@@ -19,6 +25,5 @@ export function isRoutePreloadedInStack(
   if (!navigationState || navigationState.type !== 'stack') {
     return false;
   }
-  const preloadedRoutes = (navigationState as StackNavigationState<ParamListBase>).preloadedRoutes;
-  return preloadedRoutes.some((preloaded) => preloaded.key === route.key);
+  return navigationState.routes.findIndex((item) => item.key === route.key) > navigationState.index;
 }